% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_annual_sales}
\alias{get_annual_sales}
\title{Find total annual unit sales by platform}
\usage{
get_annual_sales(df_in)
}
\arguments{
\item{df_in}{Long-format data frame of video game sales}
}
\description{
Find total annual unit sales by platform
}
