% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_shares}
\alias{get_shares}
\title{Find market share of unit sales over time}
\usage{
get_shares(df_in)
}
\arguments{
\item{df_in}{Long-format data frame of video game sales}
}
\description{
Find market share of unit sales over time
}
