% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{prep}
\alias{prep}
\title{Clean & prepare XYZ data}
\usage{
prep(df_in, how)
}
\arguments{
\item{df_in}{Data frame of XYZ data}

\item{how}{How should the data be prepared, i.e. defined by which soure's
transormation logic?}
}
\description{
Given the XYZ data, prepares it according to your predefined business logic.
}
