% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{read_sqlite}
\alias{read_sqlite}
\title{Read from the example SQLite database connection}
\usage{
read_sqlite(db, tbl)
}
\arguments{
\item{db}{SQLite database path name, passed as a string.}

\item{tbl}{Table name, passed as a string.}
}
\description{
Given a SQLite database path \code{db} and a table \code{tbl}, returns a \code{data.frame}.
}
