% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_get_annual_sales}
\alias{validate_get_annual_sales}
\title{Validate annual sales calculations}
\usage{
validate_get_annual_sales(df_in)
}
\arguments{
\item{df_in}{Data frame of calculated sales}
}
\description{
Validate annual sales calculations
}
