% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_get_shares}
\alias{validate_get_shares}
\title{Validate platform market share calculations}
\usage{
validate_get_shares(df_in)
}
\arguments{
\item{df_in}{Data frame of calculated market shares}
}
\description{
Validate platform market share calculations
}
