% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{prep}
\alias{prep}
\title{Clean & prepare XYZ data}
\usage{
prep(df_in)
}
\arguments{
\item{df_in}{Data frame of XYZ data}
}
\value{
A named \code{list} of \code{data.frame}s, where the element names serve as the
name of each table in the target database.
}
\description{
Given the XYZ data, prepares it according to your predefined business logic.
Note that this returns a named \code{list} of \code{data.frame}s.
}
