% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{write_pg}
\alias{write_pg}
\title{Write to a PostgreSQL database connection}
\usage{
write_pg(df_list, db, ...)
}
\arguments{
\item{df_list}{Data frame to write as a DB table.}

\item{db}{Database name, passed as a string.}

\item{...}{Other args passed on to \code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}.}
}
\description{
Given a database \code{db} and a table \code{tbl}, returns a \code{data.frame}.
}
